+++ worker/src/main/kotlin/com/github/k0kubun/gitstar_ranking/core/Organization.kt
@@ -3,8 +3,8 @@ package com.github.k0kubun.gitstar_ranking.core
 import java.sql.Timestamp
 
 data class Organization(
-    val id: Int,
+    val id: Long,
     val login: String,
-    val stargazersCount: Int,
+    val stargazersCount: Long,
     val updatedAt: Timestamp,
 )
