+++ worker/src/main/kotlin/com/github/k0kubun/gitstar_ranking/core/RepositoryRank.kt
@@ -1,6 +1,6 @@
 package com.github.k0kubun.gitstar_ranking.core
 
 data class RepositoryRank(
-    val stargazersCount: Int,
-    val rank: Int,
+    val stargazersCount: Long,
+    val rank: Long,
 )
