+++ worker/src/main/kotlin/com/github/k0kubun/gitstar_ranking/core/UserRank.kt
@@ -1,6 +1,6 @@
 package com.github.k0kubun.gitstar_ranking.core
 
 data class UserRank(
-    val stargazersCount: Int,
-    val rank: Int,
+    val stargazersCount: Long,
+    val rank: Long,
 )
