+++ worker/src/main/kotlin/com/github/k0kubun/gitstar_ranking/db/RepositoryRankDao.kt
@@ -9,11 +9,11 @@ import com.github.k0kubun.gitstar_ranking.core.RepositoryRank
 
 interface RepositoryRankDao {
     @SqlUpdate("delete from repository_ranks where stargazers_count between :min and :max")
-    fun deleteStarsBetween(@Bind("min") min: Int, @Bind("max") max: Int): Long
+    fun deleteStarsBetween(@Bind("min") min: Long, @Bind("max") max: Long): Long
 
     // Lower rank is larger number
     @SqlUpdate("delete from repository_ranks where rank between :highest and :lowest")
-    fun deleteRankBetween(@Bind("highest") highest: Int, @Bind("lowest") lowest: Int): Long
+    fun deleteRankBetween(@Bind("highest") highest: Long, @Bind("lowest") lowest: Long): Long
 
     @SqlBatch("insert into repository_ranks (stargazers_count, rank, created_at, updated_at) " +
         "values (:stargazersCount, :rank, current_timestamp(0), current_timestamp(0))")
