+++ worker/src/main/kotlin/com/github/k0kubun/gitstar_ranking/workers/RepositoryRankingWorker.kt
@@ -97,7 +97,7 @@ class RepositoryRankingWorker(config: GitstarRankingConfiguration) : Worker() {
         return repoRanks[repoRanks.size - 1]
     }
 
-    private fun calcProgress(child: Int, parent: Int): String {
+    private fun calcProgress(child: Long, parent: Long): String {
         return String.format("%.3f%%", child.toFloat() / parent.toFloat())
     }
 }
